Reorganizar modelos de categoría y excepción
