+++ api/API/Controllers/CategoriaController.cs
@@ -3,6 +3,7 @@ using Abstracciones.Interfaces.Flujo;
 using Abstracciones.Modelos;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Reglas.Excepciones;
 
 namespace API.Controllers
 {
